+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/versioning/semver/_internal/isTagPrefixProject.kt
@@ -1,0 +1,17 @@
+package com.javiersc.hubdle.project.extensions.config.versioning.semver._internal
+
+import com.javiersc.gradle.properties.extensions.getPropertyOrNull
+import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
+import com.javiersc.hubdle.project.extensions.config.versioning.semver.hubdleSemver
+import org.gradle.api.provider.Property
+
+internal val HubdleEnableableExtension.isTagPrefixProject: Boolean
+    get() {
+        val projectTagPrefix: Property<String> = hubdleSemver.tagPrefix
+        val tagPrefix: String = project.getPropertyOrNull("semver.tagPrefix") ?: ""
+        return when {
+            !hubdleSemver.isFullEnabled.get() -> true
+            projectTagPrefix.get() == tagPrefix -> true
+            else -> false
+        }
+    }